+++ src/components/Header.jsx
@@ -1,22 +1,26 @@
-import displayDate from '../JS/Date';
-import '../styles/Css/Header.css';
+import DisplayDate from '../JS/Date';
 const Header = () => {
-  displayDate();
   const result = (
     <header>
       <div className="fstHeader">
-        <div className="date">
-          <p id="date"></p>
-          <button type="button">العربية</button>
-          <button type="button">English</button>
+        <div>
+          {' '}
+          <div className="date">
+            <DisplayDate id="date" />
+            <button type="button">العربية</button>
+            <button type="button">English</button>
+          </div>
         </div>
-        <div className="socialLinks">
-          <img className="socialImg" src="../assets/images/facebook.svg" alt="Facebook" />
-          <img className="socialImg" src="../assets/images/twitter.svg" alt="Twitter" />
-        </div>
-        <div className="Search">
-          <input className="searchInput" type="text" />
-          <img className="socialImg" src="../assets/images/search.svg" alt="Twitter" />
+        <div>
+          {' '}
+          <div className="socialLinks">
+            <img className="socialImg" src="../assets/images/facebook.svg" alt="Facebook" />
+            <img className="socialImg" src="../assets/images/twitter.svg" alt="Twitter" />
+          </div>
+          <div className="Search">
+            <input className="searchInput" type="text" />
+            <img className="socialImg" src="../assets/images/search.svg" alt="Twitter" />
+          </div>
         </div>
       </div>
       <div className="scdHeader">
