+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
+import './styles/Css/index.css'
 import  "./assets/images/facebook.svg"
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
