+++ src/styles/Css/global.css
@@ -2,8 +2,6 @@
 * {
   margin: 0;
   padding: 0;
-  text-decoration: none;
-  outline: none;
 }
 
 /*# sourceMappingURL=global.css.map */
