+++ src/styles/Sass/App.scss
@@ -1,1 +1,5 @@
-@import 'Header.scss'
+body {
+  text-decoration: none;
+  outline: none;
+  font-size: 10px;
+}
