+++ src/styles/Sass/global.scss
@@ -13,8 +13,6 @@ $Qlpadding: calc(25% - 120px);
 * {
   margin: 0;
   padding: 0;
-  text-decoration: none;
-  outline: none;
 }
 
 @mixin columnFlex {
