+++ src/styles/Sass/index.scss
@@ -1,0 +1,3 @@
+@import 'global';
+// @import 'App.scss';
+@import 'Header.scss';
